fix for windows not implemented errors (#162)
Fixes: #150
